requirements.txt
azure-ai-inference
azure-core
python-dotenv
flask